Deprecate PQL query format
