+++ presto-pinot-toolkit/src/main/java/com/facebook/presto/pinot/PinotErrorCode.java
@@ -25,7 +25,7 @@ public enum PinotErrorCode
         implements ErrorCodeSupplier
 {
     PINOT_UNSUPPORTED_COLUMN_TYPE(0, EXTERNAL), // schema issues
-    PINOT_QUERY_GENERATOR_FAILURE(1, INTERNAL_ERROR), // Accepted a query whose pql we couldn't generate
+    PINOT_QUERY_GENERATOR_FAILURE(1, INTERNAL_ERROR), // Accepted a query whose sql we couldn't generate
     PINOT_INSUFFICIENT_SERVER_RESPONSE(2, EXTERNAL, true), // numServersResponded < numServersQueried
     PINOT_EXCEPTION(3, EXTERNAL), // Exception reported by pinot
     PINOT_HTTP_ERROR(4, EXTERNAL), // Some non okay http error code
@@ -33,7 +33,7 @@ public enum PinotErrorCode
     PINOT_UNSUPPORTED_EXPRESSION(6, INTERNAL_ERROR), // Unsupported function
     PINOT_UNABLE_TO_FIND_BROKER(7, EXTERNAL),
     PINOT_DECODE_ERROR(8, EXTERNAL),
-    PINOT_INVALID_PQL_GENERATED(9, INTERNAL_ERROR),
+    PINOT_INVALID_SQL_GENERATED(9, INTERNAL_ERROR),
     PINOT_INVALID_CONFIGURATION(10, INTERNAL_ERROR),
     PINOT_DATA_FETCH_EXCEPTION(11, EXTERNAL, true),
     PINOT_REQUEST_GENERATOR_FAILURE(12, INTERNAL_ERROR),
