+++ presto-pinot-toolkit/src/main/java/com/facebook/presto/pinot/query/PinotQueryGenerator.java
@@ -14,7 +14,6 @@
 package com.facebook.presto.pinot.query;
 
 import com.facebook.airlift.log.Logger;
-import com.facebook.presto.common.type.BigintType;
 import com.facebook.presto.common.type.FixedWidthType;
 import com.facebook.presto.common.type.JsonType;
 import com.facebook.presto.common.type.TypeManager;
@@ -59,7 +58,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.UUID;
 import java.util.stream.Collectors;
 
 import static com.facebook.presto.pinot.PinotErrorCode.PINOT_UNSUPPORTED_EXPRESSION;
@@ -136,10 +134,9 @@ public class PinotQueryGenerator
     public Optional<PinotQueryGeneratorResult> generate(PlanNode plan, ConnectorSession session)
     {
         try {
-            boolean usePinotSqlSyntax = PinotSessionProperties.isUsePinotSqlForBrokerQueries(session);
             PinotQueryGeneratorContext context = requireNonNull(plan.accept(
                             new PinotQueryPlanVisitor(session),
-                            new PinotQueryGeneratorContext(usePinotSqlSyntax)),
+                            new PinotQueryGeneratorContext()),
                     "Resulting context is null");
             return Optional.of(new PinotQueryGeneratorResult(context.toQuery(pinotConfig, session), context));
         }
@@ -149,19 +146,11 @@ public class PinotQueryGenerator
         }
     }
 
-    public enum PinotQueryFormat
-    {
-        PQL,
-        SQL
-    }
-
     public static class GeneratedPinotQuery
     {
         final String table;
         final String query;
-        final PinotQueryFormat format;
         final List<Integer> expectedColumnIndices;
-        final int groupByClauses;
         final boolean haveFilter;
         final boolean forBroker;
 
@@ -169,18 +158,14 @@ public class PinotQueryGenerator
         public GeneratedPinotQuery(
                 @JsonProperty("table") String table,
                 @JsonProperty("query") String query,
-                @JsonProperty("format") PinotQueryFormat format,
                 @JsonProperty("expectedColumnIndices") List<Integer> expectedColumnIndices,
-                @JsonProperty("groupByClauses") int groupByClauses,
                 @JsonProperty("haveFilter") boolean haveFilter,
                 @JsonProperty("forBroker") boolean forBroker)
         {
             this.table = table;
             this.query = query;
-            this.format = format;
             checkState((query != null), "Expected only one of query to be present");
             this.expectedColumnIndices = expectedColumnIndices;
-            this.groupByClauses = groupByClauses;
             this.haveFilter = haveFilter;
             this.forBroker = forBroker;
         }
@@ -197,22 +182,10 @@ public class PinotQueryGenerator
             return query;
         }
 
-        @JsonProperty("format")
-        public PinotQueryFormat getFormat()
-        {
-            return format;
-        }
-
         @JsonProperty("expectedColumnIndices")
         public List<Integer> getExpectedColumnIndices()
         {
             return expectedColumnIndices;
-        }
-
-        @JsonProperty("groupByClauses")
-        public int getGroupByClauses()
-        {
-            return groupByClauses;
         }
 
         @JsonProperty("haveFilter")
@@ -232,10 +205,8 @@ public class PinotQueryGenerator
         {
             return toStringHelper(this)
                     .add("query", query)
-                    .add("format", format)
                     .add("table", table)
                     .add("expectedColumnIndices", expectedColumnIndices)
-                    .add("groupByClauses", groupByClauses)
                     .add("haveFilter", haveFilter)
                     .add("forBroker", forBroker)
                     .toString();
@@ -247,14 +218,12 @@ public class PinotQueryGenerator
     {
         private final ConnectorSession session;
         private final boolean forbidBrokerQueries;
-        private final boolean useSqlSyntax;
         private final boolean pushdownTopnBrokerQueries;
 
         protected PinotQueryPlanVisitor(ConnectorSession session)
         {
             this.session = session;
             this.forbidBrokerQueries = PinotSessionProperties.isForbidBrokerQueries(session);
-            this.useSqlSyntax = PinotSessionProperties.isUsePinotSqlForBrokerQueries(session);
             this.pushdownTopnBrokerQueries = PinotSessionProperties.getPushdownTopnBrokerQueries(session);
         }
 
@@ -309,17 +278,14 @@ public class PinotQueryGenerator
                         new Selection(pinotExpression.getDefinition(), pinotExpression.getOrigin()));
                 newOutputs.add(variable);
             });
-            if (useSqlSyntax) {
-                // For PinotQueryGeneratorContext, selections should contain the mapping from varRef to rowExpression,
-                // and output, groupBy, orderBy objects only hold varRefs.
-                //
-                // When we try to generate Pinot query, the varRef in groupBy may not be in output.
-                // E.g. a sample Presto query: `select count(*) group by A`.
-                // - To generate PQL, we expect column `A` is always in selections.
-                // - To generate SQL, we need to hold all the mappings somewhere, which is in selections, then generate
-                //   SQL based on output object.
-                newSelections.putAll(context.getSelections());
-            }
+            // For PinotQueryGeneratorContext, selections should contain the mapping from varRef to rowExpression,
+            // and output, groupBy, orderBy objects only hold varRefs.
+            //
+            // When we try to generate Pinot query, the varRef in groupBy may not be in output.
+            // E.g. a sample Presto query: `select count(*) group by A`.
+            // - To generate SQL, we need to hold all the mappings somewhere, which is in selections, then generate
+            //   SQL based on output object.
+            newSelections.putAll(context.getSelections());
             return context.withProject(newSelections, newOutputs);
         }
 
@@ -327,8 +293,8 @@ public class PinotQueryGenerator
         public PinotQueryGeneratorContext visitTableScan(TableScanNode node, PinotQueryGeneratorContext contextIn)
         {
             PinotTableHandle tableHandle = (PinotTableHandle) node.getTable().getConnectorHandle();
-            checkSupported(!tableHandle.getPinotQuery().isPresent(), "Expect to see no existing pql");
-            checkSupported(!tableHandle.getForBroker().isPresent(), "Expect to see no existing pql");
+            checkSupported(!tableHandle.getPinotQuery().isPresent(), "Expect to see no existing sql");
+            checkSupported(!tableHandle.getForBroker().isPresent(), "Expect to see no existing sql");
             Map<VariableReferenceExpression, Selection> selections = new HashMap<>();
             LinkedHashSet<VariableReferenceExpression> outputs = new LinkedHashSet<>();
             node.getOutputVariables().forEach(outputColumn -> {
@@ -337,7 +303,7 @@ public class PinotQueryGenerator
                 selections.put(outputColumn, new Selection(pinotColumn.getColumnName(), TABLE_COLUMN));
                 outputs.add(outputColumn);
             });
-            return new PinotQueryGeneratorContext(selections, outputs, tableHandle.getTableName(), PinotSessionProperties.isUsePinotSqlForBrokerQueries(session));
+            return new PinotQueryGeneratorContext(selections, outputs, tableHandle.getTableName());
         }
 
         private String handleAggregationFunction(CallExpression aggregation, Map<VariableReferenceExpression, Selection> inputSelections)
@@ -491,7 +457,7 @@ public class PinotQueryGenerator
 
             // Make two passes over the aggregationColumnNodes: In the first pass identify all the variables that will be used
             // Then pass that context to the source
-            // And finally, in the second pass actually generate the PQL
+            // And finally, in the second pass actually generate the SQL
 
             // 1st pass
             Set<VariableReferenceExpression> variablesInAggregation = new HashSet<>();
@@ -533,7 +499,6 @@ public class PinotQueryGenerator
             LinkedHashSet<VariableReferenceExpression> groupByColumns = new LinkedHashSet<>();
             Set<VariableReferenceExpression> hiddenColumnSet = new HashSet<>(context.getHiddenColumnSet());
             int aggregations = 0;
-            boolean groupByExists = false;
 
             for (AggregationColumnNode expression : aggregationColumnNodes) {
                 switch (expression.getExpressionType()) {
@@ -546,7 +511,6 @@ public class PinotQueryGenerator
                         newSelections.put(outputColumn, new Selection(pinotColumn.getDefinition(), pinotColumn.getOrigin()));
                         groupByColumns.add(outputColumn);
                         outputs.add(outputColumn);
-                        groupByExists = true;
                         break;
                     }
                     case AGGREGATE: {
@@ -561,15 +525,6 @@ public class PinotQueryGenerator
                     default:
                         throw new PinotException(PINOT_UNSUPPORTED_EXPRESSION, Optional.empty(), "unknown aggregation expression: " + expression.getExpressionType());
                 }
-            }
-
-            // Handling non-aggregated group by
-            // E.g. `SELECT A, B FROM myTable GROUP BY A, B`
-            // In Pql mode, the generated pql is `SELECT count(*) FROM myTable GROUP BY A, B`;
-            // In Sql mode, the generated sql is still `SELECT A, B FROM myTable GROUP BY A, B`.
-            if (!useSqlSyntax && groupByExists && aggregations == 0) {
-                setHiddenField(newSelections, outputs, hiddenColumnSet);
-                aggregations++;
             }
             return context.withAggregation(newSelections, outputs, groupByColumns, aggregations, hiddenColumnSet);
         }
@@ -605,32 +560,10 @@ public class PinotQueryGenerator
             requireNonNull(context, "context is null");
             checkSupported(!forbidBrokerQueries, "Cannot push distinctLimit in segment mode");
             LinkedHashSet<VariableReferenceExpression> groupByColumns = new LinkedHashSet<>(node.getDistinctVariables());
-            if (!useSqlSyntax) {
-                // Handling PQL by adding hidden expression: count(*)
-                // E.g. `SELECT DISTINCT A, B FROM myTable LIMIT 10`
-                // In Pql mode, the generated pql is `SELECT count(*) FROM myTable GROUP BY A, B LIMIT 10`.
-                checkSupported(!context.hasAggregation(), "Aggregation already exists. Pinot doesn't support DistinctLimit with existing Aggregation");
-                checkSupported(!context.hasGroupBy(), "GroupBy already exists. Pinot doesn't support DistinctLimit with existing GroupBy");
-                Map<VariableReferenceExpression, Selection> newSelections = new HashMap<>(context.getSelections());
-                LinkedHashSet<VariableReferenceExpression> newOutputs = new LinkedHashSet<>(groupByColumns);
-                Set<VariableReferenceExpression> hiddenColumnSet = new HashSet<>();
-                setHiddenField(newSelections, newOutputs, hiddenColumnSet);
-                return context.withAggregation(newSelections, newOutputs, groupByColumns, 1, hiddenColumnSet).withLimit(node.getLimit());
-            }
             // Handle SQL by setting the groupBy columns and output columns.
             // E.g. `SELECT DISTINCT A, B FROM myTable LIMIT 10`
             // In Sql mode, the generated sql is still `SELECT A, B FROM myTable GROUP BY A, B LIMIT 10`.
             return context.withDistinctLimit(groupByColumns, node.getLimit()).withOutputColumns(node.getOutputVariables());
-        }
-
-        private void setHiddenField(Map<VariableReferenceExpression, Selection> selections,
-                LinkedHashSet<VariableReferenceExpression> outputs,
-                Set<VariableReferenceExpression> hiddenColumnSet)
-        {
-            VariableReferenceExpression hidden = new VariableReferenceExpression(Optional.empty(), UUID.randomUUID().toString(), BigintType.BIGINT);
-            selections.put(hidden, new Selection("count(*)", DERIVED));
-            outputs.add(hidden);
-            hiddenColumnSet.add(hidden);
         }
     }
 }
