+++ presto-pinot-toolkit/src/test/java/com/facebook/presto/pinot/TestPinotConfig.java
@@ -52,7 +52,6 @@ public class TestPinotConfig
                         .setInferDateTypeInSchema(true)
                         .setInferTimestampTypeInSchema(true)
                         .setForbidBrokerQueries(false)
-                        .setUsePinotSqlForBrokerQueries(true)
                         .setRestProxyServiceForQuery(null)
                         .setUseProxy(false)
                         .setGrpcHost(null)
@@ -109,7 +108,6 @@ public class TestPinotConfig
                 .put("pinot.infer-date-type-in-schema", "false")
                 .put("pinot.infer-timestamp-type-in-schema", "false")
                 .put("pinot.forbid-broker-queries", "true")
-                .put("pinot.use-pinot-sql-for-broker-queries", "false")
                 .put("pinot.rest-proxy-service-for-query", "pinot-rest-proxy-service")
                 .put("pinot.grpc-host", "localhost")
                 .put("pinot.grpc-port", "8224")
@@ -168,7 +166,6 @@ public class TestPinotConfig
                 .setInferDateTypeInSchema(false)
                 .setInferTimestampTypeInSchema(false)
                 .setForbidBrokerQueries(true)
-                .setUsePinotSqlForBrokerQueries(false)
                 .setRestProxyServiceForQuery("pinot-rest-proxy-service")
                 .setNumSegmentsPerSplit(2)
                 .setIgnoreEmptyResponses(true)
